refactor(tui): split over-budget modules under 800-LOC cap
After cargo fmt alignment, tui/src/update.rs exceeded the 800-line budget
enforced by tests/budget.rs (GH #1018). Split all three near-cap files for
comfortable headroom, following the existing file + sibling-dir pattern:

- update/mouse.rs: mouse/selection/clipboard cluster (~134 lines)
- wizard/build.rs: internal builder helpers (~110 lines)
- model/views/tests.rs: inline test module relocated (~173 lines)

Result: update.rs 666, wizard.rs 668, views.rs 570 — all well under 800.
No logic changes; `cargo test -p design-data-tui --test budget` green.
